moved user validations
